Notice: Trunk is now 3.1.3
Fixed: SMSG_INSPECT_TALENT (now it is fully implemented, on inspecting you can see all the items and their stats on the inspected player)
Fixed: Warrior charge, and some other movement-related problems (SMSG_MONSTER_MOVE MoveFlag 0x100 is now a teleport flag, so we should use 0x1000 instead; I've fixed this at multiple places, hope I haven't missed any)
Fixed: Pet Action Bar (now the icons show up on it)
Fixed: Problem where some equipped items' images showed up bad and the 3D model was not showing for them (enchantment visual's calculations overwrote them because of a bad formula)
Fixed: Shown talent points when you level up (by sending the SMSG_TALENTS_INFO packet in these cases, as it seems the old UInt32 field is not handled by the UI).
Fixed: Enchanting (I've enchanted a few items...)
Fixed: Socketing (Socketed 2 different items, one of them was a Tier 7 head piece)
Fixed: Player inspection!
Fixed: Action & Macro buttons
Fixed: WorldSession::HandleNameQueryOpcode in QueryHandler.cpp
Updated: MAX_DUNGEONS to 244+1
Fixed: Timed Quests (Thanks to Alleycat for help!)
Updated: RankTitles enum - 3.1 titles (Thanks Xuni!)
Fixed: SMSG_CREATURE_QUERY_RESPONSE, so now you see your own mount's 3D model in your pets tab (in QueryHandler.cpp, WorldSession::HandleCreatureQueryOpcode) Don't forget to clear your client's cache folder!!!
Fixed: CMSG_BUY_ITEM_IN_SLOT, so you can buy items not just by right-clicking them but also with "drag and drop" (in ItemHandler.cpp, WorldSession::HandleBuyItemInSlotOpcode)
Fixed: SMSG_ACTION_BUTTONS, first sends "Misc" then "Type" (it was swapped before)
Fixed: professions by fixing SMSG_NAME_QUERY_RESPONSE, as Andy told me this is why the professions couldn't be used so far
Fixed: SMSG_LOGIN_SETTIMESPEED packet, so now the client receives the realm time properly again (it started the clock from 0:00 before)
Fixed: the quest symbols, as Blizzard introduced 2 new ones, and had to shift the old ones up by 2 because of this
Updated: various Battleground/Arena-related packets.
Updated: A9 code
Fixed: Transports
Fixed: Gameobject rotation, so the road signs, beds, etc. now face in their intended direction.
Fixed: GAMEOBJECT_PARENTROTATION_0x field handling (in UpdateFields.h, and everywhere else where I commented them out before).
Fixed: Scripts compilation
Fixed: Npc movement - Now they run towards you on aggro and not just "instant teleport" there.
Fixed: Taxis - You can fly on your birdies now :)
Fixed: Bags & Bank Tabs

Thanks to Mangos and Aspire for some inspiration.

+++ src/arcemu-world/QuestMgr.h
@@ -105,7 +105,7 @@ public:
 	ARCEMU_INLINE int32 QuestHasMob(Quest* qst, uint32 mob)
 	{
 		for(uint32 i = 0; i < 4; ++i)
-			if(qst->required_mob[i] == mob)
+			if(qst->required_mob[i] == (int32)mob)
 				return qst->required_mobcount[i];
 		return -1;
 	}
@@ -113,7 +113,7 @@ public:
 	ARCEMU_INLINE int32 GetOffsetForMob(Quest *qst, uint32 mob)
 	{
 		for(uint32 i = 0; i < 4; ++i)
-			if(qst->required_mob[i] == mob)
+			if(qst->required_mob[i] == (int32)mob)
 				return i;
 
 		return -1;
